apps/video-intelligence/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.10.0
pillow>=10.1.0
pymongo>=4.6.0
voyageai==0.3.7
openai>=1.3.8
python-dotenv>=1.0.0
websockets>=12.0
numpy>=1.26.0
aiofiles>=23.2.1
anthropic>=0.90,<1.0
langgraph==1.1.5
langgraph-prebuilt==1.0.9
langgraph-checkpoint-mongodb==0.3.1
langgraph-store-mongodb==0.2.0
langchain-core>=0.3,<1.0
